Pass alpha and beta as float

+++ include/ctranslate2/primitives/cpu_mkl.h
@@ -60,7 +60,7 @@ namespace ctranslate2 {
   void primitives<Device::CPU>::gemm(const int16_t* a, const int16_t* b,
                                      bool transpose_a, bool transpose_b,
                                      size_t m, size_t n, size_t k,
-                                     int16_t alpha, int32_t beta,
+                                     float alpha, float beta,
                                      int32_t* c);
 
   template<>
